feat: Validation added and Spectre console package is used for tabular format display

+++ src/Inventory Management/Program.cs
@@ -8,43 +8,7 @@ namespace Assignments
     /// </summary>
     internal class Program
     {
-        private static InventoryServices _service = new InventoryServices();
-
-        /// <summary>
-        /// InventoryMenu is a enum.
-        /// </summary>
-        internal enum InventoryMenu
-        {
-            /// <summary>
-            /// AddProduct is a enum value assigned with value 1.
-            /// </summary>
-            AddProduct = 1,
-
-            /// <summary>
-            /// ViewProduct is a enum value assigned with value 2.
-            /// </summary>
-            ViewProducts = 2,
-
-            /// <summary>
-            /// SearchProduct is a enum value assigned with value 3.
-            /// </summary>
-            SearchProduct = 3,
-
-            /// <summary>
-            /// UpdateProduct is a enum value assigned with value 4.
-            /// </summary>
-            UpdateProduct = 4,
-
-            /// <summary>
-            /// DeleteProduct is a enum value assigned with value 5.
-            /// </summary>
-            DeleteProduct = 5,
-
-            /// <summary>
-            /// Exit is a enum value assigned with value 6.
-            /// </summary>
-            Exit = 6,
-        }
+        private static readonly InventoryServices _service = new InventoryServices();
 
         /// <summary>
         /// Main method is the entry point of the program.
@@ -69,39 +33,39 @@ Enter Choice:";
 
                     switch (choice)
                     {
-                        case (int)InventoryMenu.AddProduct:
+                        case (int)InventoryModel.InventoryMenu.AddProduct:
                             AddProduct();
                             break;
 
-                        case (int)InventoryMenu.ViewProducts:
+                        case (int)InventoryModel.InventoryMenu.ViewProducts:
                             ViewProducts();
                             break;
 
-                        case (int)InventoryMenu.SearchProduct:
+                        case (int)InventoryModel.InventoryMenu.SearchProduct:
                             SearchProduct();
                             break;
 
-                        case (int)InventoryMenu.UpdateProduct:
+                        case (int)InventoryModel.InventoryMenu.UpdateProduct:
                             UpdateProduct();
                             break;
 
-                        case (int)InventoryMenu.DeleteProduct:
+                        case (int)InventoryModel.InventoryMenu.DeleteProduct:
                             DeleteProduct();
                             break;
 
-                        case (int)InventoryMenu.Exit:
-                            Helper.WriteColored("Exiting...", ConsoleColor.Red);
+                        case (int)InventoryModel.InventoryMenu.Exit:
+                            Helper.WriteColored("Exiting...", ConsoleColor.Cyan);
                             Thread.Sleep(1000);
                             return;
 
                         default:
-                            Console.WriteLine("Invalid Choice.");
+                            Helper.WriteColored("Invalid Choice.", ConsoleColor.Red);
                             break;
                     }
                 }
                 catch (FormatException)
                 {
-                    Console.WriteLine("Invalid Input. Please enter numeric value.");
+                    Helper.WriteColored("Invalid Input. Please enter numeric value.", ConsoleColor.Red);
                 }
                 catch (Exception ex)
                 {
@@ -112,21 +76,33 @@ Enter Choice:";
 
         private static void DeleteProduct()
         {
+            if (_service.InventoryIsEmpty())
+            {
+                Helper.WriteColored("Inventory is empty.", ConsoleColor.Red);
+                return;
+            }
+
             Console.Write("Enter Product ID: ");
             int deleteId = Convert.ToInt32(Console.ReadLine());
 
             if (_service.DeleteProduct(deleteId))
             {
-                Console.WriteLine("Product Deleted.");
+                Helper.WriteColored("Product Deleted.", ConsoleColor.Green);
             }
             else
             {
-                Console.WriteLine("Product Not Found.");
+                Helper.WriteColored("Product Not Found.", ConsoleColor.Red);
             }
         }
 
         private static void UpdateProduct()
         {
+            if (_service.InventoryIsEmpty())
+            {
+                Helper.WriteColored("Inventory is empty.", ConsoleColor.Red);
+                return;
+            }
+
             InventoryModel update = new InventoryModel();
 
             Console.Write("Enter Product ID : ");
@@ -143,22 +119,28 @@ Enter Choice:";
 
             if (_service.UpdateProduct(update))
             {
-                Console.WriteLine("Product Updated.");
+                Helper.WriteColored("Product Updated.", ConsoleColor.Green);
             }
             else
             {
-                Console.WriteLine("Product Not Found.");
+                Helper.WriteColored("Product Not Found.", ConsoleColor.Red);
             }
         }
 
         private static void SearchProduct()
         {
+            if (_service.InventoryIsEmpty())
+            {
+                Helper.WriteColored("Inventory is empty.", ConsoleColor.Red);
+                return;
+            }
+
             Console.Write("Enter Product Name or ID: ");
             string? input = Console.ReadLine()?.Trim();
 
             if (string.IsNullOrWhiteSpace(input))
             {
-                Console.WriteLine("Invalid input. Please enter a product name or ID.");
+                Helper.WriteColored("Invalid input. Please enter a product name or ID.", ConsoleColor.Red);
                 return;
             }
 
@@ -168,7 +150,7 @@ Enter Choice:";
 
                 if (item == null)
                 {
-                    Console.WriteLine("Product Not Found.");
+                    Helper.WriteColored("Product Not Found.", ConsoleColor.Red);
                 }
                 else
                 {
@@ -182,7 +164,7 @@ Enter Choice:";
 
                 if (results == null || results.Count == 0)
                 {
-                    Console.WriteLine("No Product Found.");
+                    Helper.WriteColored("No Products Found.", ConsoleColor.Red);
                 }
                 else
                 {
@@ -197,20 +179,21 @@ Enter Choice:";
 
         private static void ViewProducts()
         {
+            if (_service.InventoryIsEmpty())
+            {
+                Helper.WriteColored("Inventory is empty.", ConsoleColor.Red);
+                return;
+            }
+
             var products = _service.GetProducts();
 
             if (products.Count == 0)
             {
-                Console.WriteLine("No Products Available.");
+                Helper.WriteColored("No Products Available.", ConsoleColor.Red);
             }
             else
             {
-                Console.WriteLine("ID\tName\tPrice\tQuantity");
-
-                foreach (var p in products)
-                {
-                    Console.WriteLine($"{p.ProductID}\t{p.ProductName}\t{p.Price}\t{p.Quantity}");
-                }
+                Helper.PrintingTable(products);
             }
         }
 
@@ -230,13 +213,19 @@ Enter Choice:";
             Console.Write("Quantity: ");
             product.Quantity = Convert.ToInt32(Console.ReadLine());
 
+            if (!Helper.ValidateData(product))
+            {
+                Helper.WriteColored("The data entered is not valid.", ConsoleColor.Red);
+                return;
+            }
+
             if (_service.AddProduct(product))
             {
-                Console.WriteLine("Product Added Successfully.");
+                Helper.WriteColored("Product Added Successfully.", ConsoleColor.Green);
             }
             else
             {
-                Console.WriteLine("Product ID already exists.");
+                Helper.WriteColored("Product ID already exists.", ConsoleColor.Red);
             }
         }
     }
